adding changes to cleanbuildmachines
git-svn-id: svn+ssh://svn.cern.ch/reps/lcgsoft/trunk/lcgcmake@7134 4525493e-7705-40b1-a816-d608a930855b

+++ cdash/cleanbuildmachines.py
@@ -4,25 +4,25 @@ import sys, shutil, os
 from os import path
 from datetime import datetime, timedelta
 
-two_days_ago = datetime.now() - timedelta(days=2)
-check_file = "controlfile"
-work_directory = os.environ["WORKSPACE"]
+two_days_ago = datetime.now() - timedelta(hours=2)
+work_directory = "/build/jenkins/workspace"
 
-if os.path.exists(check_file):
-    print  "the controlfile exists"
-else:
-    print "controlfile does not exist, we stop here"
-    sys.exit(-1)
+for dirs in os.listdir(work_directory):
+    if os.path.isfile(dirs):
+        pass
+    else:
+        fullpath = work_directory+"/"+dirs
+        check_file = fullpath+"/controlfile"
 
-filetime = datetime.fromtimestamp(path.getctime(check_file))
-
-
-if filetime < two_days_ago:
-  print "controlFile is more than two days old"
-  for root, dirs, files in os.walk(work_directory):
-      for f in files:
-          os.unlink(os.path.join(root, f))
-      for d in dirs:
-          shutil.rmtree(os.path.join(root, d))
-else:
-  print "controlFile is less than two days old. Nothing to remove"  
+        if os.path.exists(check_file):
+            filetime = datetime.fromtimestamp(path.getctime(check_file))
+            print "the controlfile exists at:  ", fullpath
+            if filetime < two_days_ago:
+                print "controlFile is more than two days old"
+                for root, dirs, files in os.walk(fullpath):
+                    for f in files:
+                        os.unlink(os.path.join(root, f))
+                    for d in dirs:
+                        shutil.rmtree(os.path.join(root, d))
+            else:
+                print "controlFile is less than two days old. Nothing to remove"
